# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/usermanagement/data": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "Display of relation user and role",
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "Display of relation user and role",
                "parameters": [
                    {
                        "description": "Dados para atualizar a role do usuário",
                        "name": "usermanagement",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usermanagementdto.UpdateRelationUserRole"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "usermanagementdto.UpdateRelationUserRole": {
            "type": "object",
            "properties": {
                "id_role": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                }
            }
        }
    }
}